+++ sticknotes-frontend/src/app/board-container/board-container.component.html
@@ -12,7 +12,7 @@
                 <mat-icon>edit</mat-icon>
               </button>
             </div>
-            <h3>Created {{sidenavBoardData.creationDate | amCalendar}}</h3>
+            <h3>Created {{getBoardCreatedDate() | amCalendar}}</h3>
           </mat-card-content>
         </mat-card>
       </div>
