+++ sticknotes-frontend/src/app/new-note/new-note.component.html
@@ -2,7 +2,12 @@
   <form [formGroup]="newNoteForm" class="new-note-form">
     <mat-form-field>
       <mat-label>Note</mat-label>
-      <input (keydown.enter)="handleSubmit()" formControlName="content" matInput placeholder="What would you like to note?">
+      <textarea matInput formControlName="content" (keydown.enter)="handleSubmit()" cdkTextareaAutosize
+        placeholder="What would you like to note?" #autosize="cdkTextareaAutosize" cdkAutosizeMinRows="1"
+        cdkAutosizeMaxRows="5"></textarea>
+      <mat-error *ngIf="!this.newNoteForm.valid">
+        Please enter a valid note
+      </mat-error>
     </mat-form-field>
     <label id="colorpicker-label">Note color</label>
     <mat-radio-group aria-labelledby="colorpicker-label" class="color-picker" formControlName="options">
@@ -10,5 +15,6 @@
           class="circle" [style]="'background-color:' + getColorHexValue((i + 1).toString())"></span></mat-radio-button>
     </mat-radio-group>
   </form>
-  <button mat-raised-button color="primary" (click)="handleSubmit()" [disabled]="sendingData">{{submitButtonText}}</button>
+  <button mat-raised-button color="primary" (click)="handleSubmit()"
+    [disabled]="sendingData">{{submitButtonText}}</button>
 </div>
