+++ sticknotes-frontend/src/app/services/notes-api.service.ts
@@ -16,7 +16,7 @@ export class NotesApiService {
     // use mocked response for now
     const mockedNoteResponse: Note = {
       key: 'key',
-      creationDate: new Date().toISOString(),
+      creationDate: this.currentTimestamp,
       creator: 'googler@google.com',
       x: note.x,
       y: note.y,
@@ -36,5 +36,9 @@ export class NotesApiService {
   public deleteNote(noteKey: string): Observable<void> {
     // return this.http.delete<void>(`api/notes?key=${noteKey}`);
     return of(undefined);
+  }
+
+  get currentTimestamp(): string {
+    return Math.round(new Date().getTime()).toString()
   }
 }
